commented out set clipboard command
This was causing me problems: If I copied some text into my
system clipboard (say in the browser) then went to a Vim
session I had going because I wanted to paste this text in.

Then, suppose that before pasting in the text (it's now in
the * buffer), I were to cut or copy something to the
unnamed register, it was also putting that into the *
buffer and causing me to lose what was there.